Sub info GET

+++ src/main/java/grad/service/CoreService.java
@@ -397,8 +397,8 @@ public class CoreService {
                 if (eventType.equals(MessageUtil.EVENT_TYPE_SUBSCRIBE)) {
                     JSONObject User_Info = UserManager.getUser_Info(fromUserName);
                     String Subscriber_Sex = "未知";
-                    if((int)User_Info.get("sex") == 1)  {Subscriber_Sex = "男";}
-                    else if ((int)User_Info.get("sex") == 2) {Subscriber_Sex = "女";}
+                    if(User_Info.get("sex") == 1)  {Subscriber_Sex = "男";}
+                    else if (User_Info.get("sex") == 2) {Subscriber_Sex = "女";}
                     Subscriber_Info new_subscriber_info = new Subscriber_Info(
                             (String)User_Info.get("openid"), (String)User_Info.get("nickname"), Subscriber_Sex, (String)User_Info.get("language"),
                             (String)User_Info.get("city"), (String)User_Info.get("province"), (String)User_Info.get("country"), (String)User_Info.get("headimgurl"),
